Isolate `issue` commands
