+++ pkg/cmd/issue/reopen/reopen.go
@@ -1,4 +1,4 @@
-package close
+package reopen
 
 import (
 	"fmt"
@@ -7,14 +7,14 @@ import (
 	"github.com/cli/cli/api"
 	"github.com/cli/cli/internal/config"
 	"github.com/cli/cli/internal/ghrepo"
-	"github.com/cli/cli/pkg/cmd/pr/shared"
+	"github.com/cli/cli/pkg/cmd/issue/shared"
 	"github.com/cli/cli/pkg/cmdutil"
 	"github.com/cli/cli/pkg/iostreams"
 	"github.com/cli/cli/utils"
 	"github.com/spf13/cobra"
 )
 
-type CloseOptions struct {
+type ReopenOptions struct {
 	HttpClient func() (*http.Client, error)
 	Config     func() (config.Config, error)
 	IO         *iostreams.IOStreams
@@ -23,16 +23,16 @@ type CloseOptions struct {
 	SelectorArg string
 }
 
-func NewCmdClose(f *cmdutil.Factory, runF func(*CloseOptions) error) *cobra.Command {
-	opts := &CloseOptions{
+func NewCmdReopen(f *cmdutil.Factory, runF func(*ReopenOptions) error) *cobra.Command {
+	opts := &ReopenOptions{
 		IO:         f.IOStreams,
 		HttpClient: f.HttpClient,
 		Config:     f.Config,
 	}
 
 	cmd := &cobra.Command{
-		Use:   "close {<number> | <url> | <branch>}",
-		Short: "Close a pull request",
+		Use:   "reopen {<number> | <url>}",
+		Short: "Reopen issue",
 		Args:  cobra.ExactArgs(1),
 		RunE: func(cmd *cobra.Command, args []string) error {
 			// support `-R, --repo` override
@@ -45,39 +45,36 @@ func NewCmdClose(f *cmdutil.Factory, runF func(*CloseOptions) error) *cobra.Comm
 			if runF != nil {
 				return runF(opts)
 			}
-			return closeRun(opts)
+			return reopenRun(opts)
 		},
 	}
 
 	return cmd
 }
 
-func closeRun(opts *CloseOptions) error {
+func reopenRun(opts *ReopenOptions) error {
 	httpClient, err := opts.HttpClient()
 	if err != nil {
 		return err
 	}
 	apiClient := api.NewClientFromHTTP(httpClient)
 
-	pr, baseRepo, err := shared.PRFromArgs(apiClient, opts.BaseRepo, nil, nil, opts.SelectorArg)
+	issue, baseRepo, err := shared.IssueFromArg(apiClient, opts.BaseRepo, opts.SelectorArg)
 	if err != nil {
 		return err
 	}
 
-	if pr.State == "MERGED" {
-		fmt.Fprintf(opts.IO.ErrOut, "%s Pull request #%d (%s) can't be closed because it was already merged", utils.Red("!"), pr.Number, pr.Title)
-		return cmdutil.SilentError
-	} else if pr.Closed {
-		fmt.Fprintf(opts.IO.ErrOut, "%s Pull request #%d (%s) is already closed\n", utils.Yellow("!"), pr.Number, pr.Title)
+	if !issue.Closed {
+		fmt.Fprintf(opts.IO.ErrOut, "%s Issue #%d (%s) is already open\n", utils.Yellow("!"), issue.Number, issue.Title)
 		return nil
 	}
 
-	err = api.PullRequestClose(apiClient, baseRepo, pr)
+	err = api.IssueReopen(apiClient, baseRepo, *issue)
 	if err != nil {
-		return fmt.Errorf("API call failed: %w", err)
+		return err
 	}
 
-	fmt.Fprintf(opts.IO.ErrOut, "%s Closed pull request #%d (%s)\n", utils.Red("✔"), pr.Number, pr.Title)
+	fmt.Fprintf(opts.IO.ErrOut, "%s Reopened issue #%d (%s)\n", utils.Green("✔"), issue.Number, issue.Title)
 
 	return nil
 }
