+++ pkg/cmd/issue/view/fixtures/issueView_preview.json
