+++ pkg/cmd/issue/view/fixtures/issueView_previewClosedState.json
