+++ pkg/cmd/issue/view/fixtures/issueView_previewWithEmptyBody.json
