+++ pkg/cmd/issue/view/fixtures/issueView_previewWithMetadata.json
